Change how we handle the FileTime inside the FileInformation Struct
Instead of instancing 3 local variable in the initialize method
which never be use we rely on FFI to correctly instanciate them.

This change has a side effect of reducing the memory on the Heap to keep
track of the native memory and make the garbage collector work a bit
smoother on the values.